feat(redux): react-redux, redux, reduxjs/toolkit
* add RootNavigator
* add redux system
* add userSlice
