+++ app/navigation/index.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+
+import {LoginScreen, FeedsScreen} from '../screens';
+
+const Stack = createNativeStackNavigator();
+
+export const RootNavigator = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="Login" component={LoginScreen} />
+        <Stack.Screen name="Feeds" component={FeedsScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
