+++ app/redux/index.ts
@@ -1,0 +1,7 @@
+import {configureStore} from '@reduxjs/toolkit';
+
+import {userSlice} from './user';
+
+export const store = configureStore({
+  reducer: userSlice.reducer,
+});
