+++ app/redux/user.ts
@@ -1,0 +1,15 @@
+import {createSlice} from '@reduxjs/toolkit';
+
+export const userSlice = createSlice({
+  name: 'user',
+  initialState: {
+    email: null,
+  },
+  reducers: {
+    setEmail: (state, action) => {
+      state.email = action.payload;
+    },
+  },
+});
+
+export const {setEmail} = userSlice.actions;
